osint-framework/requirements.txt
# Core dependencies
aiohttp>=3.8.0
asyncio-throttle>=1.0.2
beautifulsoup4>=4.11.0
blake3>=0.3.1
cryptography>=3.4.8
dataclasses>=0.6; python_version<"3.7"
fastapi>=0.85.0
jinja2>=3.1.0
lxml>=4.9.0
pydantic>=1.10.0
python-multipart>=0.0.5
redis>=4.3.0
sqlalchemy>=1.4.0
uvicorn>=0.18.0
pyyaml>=6.0
pytest>=7.0.0
pytest-asyncio>=0.20.0
pytest-cov>=4.0.0
httpx>=0.23.0
python-dotenv>=0.19.0

# Enhanced capabilities for advanced OSINT
scikit-learn>=1.3.0
pandas>=1.5.0
numpy>=1.24.0
nltk>=3.8.0
spacy>=3.6.0
textdistance>=4.5.0
phonetics>=0.4.0
python-whois>=0.8.0
dnspython>=2.3.0
cryptography>=3.4.8
pillow>=9.5.0
opencv-python>=4.8.0
face-recognition>=1.3.0
selenium>=4.15.0
undetected-chromedriver>=3.5.0
playwright>=1.40.0
requests>=2.31.0
urllib3>=1.26.0
certifi>=2023.7.0

# Advanced data processing
elasticsearch>=8.10.0
psycopg2-binary>=2.9.0
alembic>=1.11.0
celery>=5.3.0
flower>=2.0.0

# Machine learning and NLP
transformers>=4.30.0
torch>=2.0.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
sentencepiece>=0.1.99
tokenizers>=0.13.0

# Security and compliance
passlib>=1.7.4
bcrypt>=4.0.1
python-jose>=3.3.0
python-multipart>=0.0.5
itsdangerous>=2.1.0

# Monitoring and observability
prometheus-client>=0.17.0
grafana-api>=1.0.3
sentry-sdk>=1.29.0
structlog>=23.1.0
rich>=13.4.0
typer>=0.9.0

# Enhanced connectors
tweepy>=4.14.0
linkedin-api>=0.1.5
PyGithub>=1.59.0
google-api-python-client>=2.100.0
bing-web-search>=7.2.0
duckduckgo-search>=3.8.0
waybackpy>=3.0.0
shodan>=1.29.0
censys>=2.2.0

# Data validation and quality
pydantic-settings>=2.0.0
email-validator>=2.0.0
phonenumbers>=8.13.0
python-dateutil>=2.8.0
pytz>=2023.3

# Advanced reporting
weasyprint>=59.0
reportlab>=4.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
jupyter>=1.0.0

# Testing and development
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0
pre-commit>=3.3.0
coverage>=7.3.0
structlog>=22.1.0
